requirements.txt
Gooey==0.9.2.3
Pillow==3.3.1
# wxPython==2.8.12.1 # install manually
